+++ src/constants/settings.ts
@@ -1,43 +1,23 @@
 export type SettingsStorage = {
-  version: number
+  version: 1.1
   lists: {
     common: boolean
     fakeNews: boolean
     gambling: boolean
     social: boolean
+    ipGrabbers: boolean
   }
 }
 
 export const DEFAULT_SETTINGS: SettingsStorage = {
-  version: 1,
+  version: 1.1,
   lists: {
     common: true,
-    fakeNews: true,
-    gambling: true,
-    social: true,
+    fakeNews: false,
+    gambling: false,
+    social: false,
+    ipGrabbers: false,
   },
 }
 
 export const LATEST_SETTINGS_VERSION = DEFAULT_SETTINGS.version
-
-export enum SettingsConn {
-  reload = 'type.reload',
-}
-
-export enum PopupConn {
-  // Get the ads blocked
-  getAds = 'type.getAds',
-  returnAds = 'type.returnAds',
-
-  // Whitelist management
-  addWhitelist = 'type.addWhitelist',
-  removeWhitelist = 'type.removeWhitelist',
-  getWhitelist = 'type.getWhitelist',
-  returnWhitelist = 'type.returnWhitelist',
-}
-
-export enum StatsConn {
-  // Get LT data
-  getLT = 'type.getLT',
-  returnLT = 'type.returnLT',
-}