+++ src/frontend/ui/settings/app.tsx
@@ -3,14 +3,9 @@
 import React, { Component } from 'react'
 import { remoteFn } from '../../../backend/lib/remoteFunctions'
 
-import { DEFAULT_SETTINGS, SettingsConn } from '../../../constants/settings'
+import { DEFAULT_SETTINGS, SettingsStorage } from '../../../constants/settings'
 import { Button, Checkbox } from '../common'
 import styles from './settings.module.css'
-
-// Connect to the background script
-const backgroundScript = browser.runtime.connect({
-  name: 'co.dothq.shield.ui.settings',
-})
 
 interface AppState {
   settings?: any
@@ -45,7 +40,7 @@ class SettingsApp extends Component<AppState> {
   }
 
   render() {
-    let settings = this.state.settings
+    let settings: SettingsStorage = this.state.settings
 
     return (
       <div className={styles.page}>
@@ -81,6 +76,15 @@ class SettingsApp extends Component<AppState> {
                 }}
               >
                 Social media filter list
+              </Checkbox>
+              <Checkbox
+                value={settings.lists.ipGrabbers}
+                onChange={() => {
+                  settings.lists.ipGrabbers = !settings.lists.ipGrabbers
+                  this.setState({ settings })
+                }}
+              >
+                IP Grabbers filter list
               </Checkbox>
             </div>